enable gemv
